Add missing criu symbols to criu_unsupported.go
[NO NEW TESTS NEEDED]
