+++ pkg/criu/criu_unsupported.go
@@ -3,6 +3,14 @@
 
 package criu
 
+func CheckForCriu(version int) bool {
+	return false
+}
+
 func MemTrack() bool {
 	return false
+}
+
+func GetCriuVestion() (int, error) {
+	return MinCriuVersion, nil
 }
